Move the history stuff into a concern
